Moved the handlers specific to the portal into their own file
