+++ ceno-client/src/portal.go
@@ -4,12 +4,56 @@ import (
 	"bytes"
 	"encoding/base64"
 	"encoding/json"
+	"fmt"
 	"github.com/nicksnyder/go-i18n/i18n"
 	"html/template"
 	"net/http"
 	"os"
 	"path"
 )
+
+type PortalPath struct {
+	PageName string
+	Href     string
+}
+
+// Location of the JSON file containing the merged translated strings
+var allJSONPath string = path.Join(".", "locale", "all.json")
+
+/**
+ * ceno-client/locale/all.json contains data formatted like
+{
+ 	"en": {
+		"string1": "content content content"
+	},
+	"fr": {
+		"string1": "french french french"
+	}
+}
+*/
+type IdentifiedString struct {
+	Identifier string
+	Content    string
+}
+
+type LanguageStrings struct {
+	Name    string
+	Locale  string
+	Strings []IdentifiedString
+}
+
+type LanguageStringJSON map[string]map[string]string
+
+/**
+ * JSON files containing information about articles stored in the distributed cache (Freenet)
+ * are named like `json-files/<base64(feed's url)>.json`
+ * @param feedUrl - The URL of the RSS/Atom feed to retrieve information about articles from
+ * @return the path to the article of interest's respective JSON file on disk
+ */
+func articlesFilename(feedUrl string) string {
+	b64FeedUrl := base64.StdEncoding.EncodeToString([]byte(feedUrl))
+	return path.Join(".", "json-files", b64FeedUrl+".json")
+}
 
 /**
  * Get information about feeds to be injected into the portal page.
@@ -52,37 +96,148 @@ func InitModuleWithFeeds() (map[string]interface{}, error) {
 }
 
 /**
- * Build the portal page with information about articles already inserted into Freenet
+ * Get information about articles from a given feed to be injected into the portal page.
+ * @param {string} feedUrl - The URL of the feed to fetch articles from
+ * @return a map with a "feeds" key and corresponding array of Feed structs and an optional error
  */
-func CreatePortalPage(w http.ResponseWriter, r *http.Request) {
-	T, _ := i18n.Tfunc(os.Getenv(LANG_ENVVAR), DEFAULT_LANG)
-	t, _ := template.ParseFiles(path.Join(".", "views", "feeds.html"))
-	moduleData, feedsErr := InitModuleWithFeeds()
-	if feedsErr != nil {
-		// We could end up with a decode error here, but it's not quite practical to ditinguish.
-		HandleCCError(ERR_NO_FEEDS_FILE, feedsErr.Error(), ErrorState{
-			"responseWriter": w,
-			"request":        r,
-		})
-		return
+func InitModuleWithArticles(feedUrl string) (map[string]interface{}, error) {
+	articleInfo := ArticleInfo{}
+	var decodeErr error
+	result := Lookup(feedUrl)
+	if result.Complete && result.Found {
+		// Serve whatever the LCS gave us as the most recent articles list for
+		// the feed we want to see.
+		decoder := json.NewDecoder(bytes.NewReader([]byte(result.Bundle)))
+		decodeErr = decoder.Decode(&articleInfo)
+	} else {
+		// Before the first complete lookup, serve from the files
+		// distributed with the client.
+		articleInfoFile, openErr := os.Open(articlesFilename(feedUrl))
+		if openErr != nil {
+			return nil, openErr
+		}
+		defer articleInfoFile.Close()
+		decoder := json.NewDecoder(articleInfoFile)
+		decodeErr = decoder.Decode(&articleInfo)
 	}
-	moduleData["articles"] = T("articles_word")
-	moduleData["lastPublished"] = T("last_published_word")
-	moduleData["latest"] = T("latest_word")
-	moduleDataMarshalled, err := json.Marshal(moduleData)
-	var module string
+	if decodeErr != nil {
+		return nil, decodeErr
+	}
+	mapping := make(map[string]interface{})
+	mapping["Articles"] = articleInfo.Items
+	mapping["Version"] = articleInfo.Version
+	return mapping, nil
+}
+
+func stringifyLanguages(langStrings LanguageStringJSON) string {
+	asBytes, _ := json.Marshal(langStrings)
+	return string(asBytes)
+}
+
+func loadLanguageStrings() ([]LanguageStrings, LanguageStringJSON, error) {
+	// Dear Glob.
+	langStrings := make(LanguageStringJSON)
+	decodedStrings := []LanguageStrings{}
+	file, err := os.Open(allJSONPath)
 	if err != nil {
-		HandleCCError(ERR_CORRUPT_JSON, err.Error(), ErrorState{
-			"responseWriter": w,
-			"request":        r,
-		})
-		return
+		return decodedStrings, langStrings, err
 	}
-	module = string(moduleDataMarshalled[:])
-	t.Execute(w, map[string]interface{}{
-		"Previous":         T("previous_word"),
-		"More":             T("more_word"),
-		"PortalBlurb":      T("portal_blurb"),
-		"CenoPortalModule": module,
-	})
+	defer file.Close()
+	decoder := json.NewDecoder(file)
+	decodeErr := decoder.Decode(&langStrings)
+	if decodeErr != nil {
+		return decodedStrings, langStrings, decodeErr
+	}
+	// Use the configuration as a guide to explore the merged languages json file and construct
+	// structures containing all the information relevant to the portal page about text.
+	for _, availableLanguage := range Configuration.PortalLanguages {
+		stringPairs, found := langStrings[availableLanguage.Locale]
+		if !found {
+			continue
+		}
+		languageStrings := LanguageStrings{}
+		languageStrings.Name = availableLanguage.Name
+		languageStrings.Locale = availableLanguage.Locale
+		for identifier, content := range stringPairs {
+			languageStrings.Strings = append(languageStrings.Strings, IdentifiedString{identifier, content})
+		}
+		decodedStrings = append(decodedStrings, languageStrings)
+	}
+	return decodedStrings, langStrings, nil
+}
+
+func PortalIndexHandler(w http.ResponseWriter, r *http.Request) {
+	fmt.Println("Got request for test portal page")
+	t, _ := template.ParseFiles("./views/index.html", "./views/nav.html", "./views/resources.html", "./views/scripts.html")
+	module := map[string]interface{}{}
+	languageStrings, langStringsJson, readErr := loadLanguageStrings()
+	if readErr != nil {
+		fmt.Println("Error loading language")
+		fmt.Println(readErr)
+	} else {
+		// For the language selection menu
+		module["LanguageStrings"] = languageStrings
+		// For the javascript code that applies strings
+		module["LanguageStringsAsJSON"] = stringifyLanguages(langStringsJson)
+	}
+	t.Execute(w, module)
+}
+
+func PortalChannelsHandler(w http.ResponseWriter, r *http.Request) {
+	fmt.Println("Got request for test channels page")
+	t, _ := template.ParseFiles("./views/channels.html", "./views/nav.html", "./views/resources.html", "./views/breadcrumbs.html", "./views/scripts.html")
+	module, err := InitModuleWithFeeds()
+	if err != nil {
+		fmt.Println("Error loading feeds")
+		fmt.Println(err)
+		t.Execute(w, nil)
+	} else {
+		module["Breadcrumbs"] = []PortalPath{
+			{"CeNO", "/testportal"},
+			{"Channel Selector", "/testchannels"},
+		}
+		languageStrings, langStringsJson, readErr := loadLanguageStrings()
+		if readErr != nil {
+			fmt.Println("Error loading languages")
+			fmt.Println(err)
+		} else {
+			// For the language selection menu
+			module["LanguageStrings"] = languageStrings
+			// For the javascript code that applies strings
+			module["LanguageStringsAsJSON"] = stringifyLanguages(langStringsJson)
+		}
+		t.Execute(w, module)
+	}
+}
+
+func PortalArticlesHandler(w http.ResponseWriter, r *http.Request) {
+	T, _ := i18n.Tfunc(os.Getenv(LANG_ENVVAR), DEFAULT_LANG)
+	t, _ := template.ParseFiles("./views/articles.html", "./views/nav.html", "./views/resources.html", "./views/breadcrumbs.html", "./views/scripts.html")
+	// TODO - Get the feed from the URL
+	module, err := InitModuleWithArticles("https://news.ycombinator.com/rss")
+	if err != nil {
+		fmt.Println("Error loading articles")
+		fmt.Println(err)
+		t.Execute(w, nil)
+	} else {
+		module["PublishedWord"] = T("published_word")
+		module["AuthorWord"] = T("authors_word")
+		module["Title"] = "Testing 1 2 3"
+		module["Breadcrumbs"] = []PortalPath{
+			{"CeNO", "/testportal"},
+			{"Channel Selector", "/testchannels"},
+			{"Hacker News", "/testarticles"},
+		}
+		languageStrings, langStringsJson, readErr := loadLanguageStrings()
+		if readErr != nil {
+			fmt.Println("Error loading languages")
+			fmt.Println(err)
+		} else {
+			// For the language selection menu
+			module["LanguageStrings"] = languageStrings
+			// For the javascript code that applies strings
+			module["LanguageStringsAsJSON"] = stringifyLanguages(langStringsJson)
+		}
+		t.Execute(w, module)
+	}
 }
